# server/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Bakery Management System API",
      "description": "A sample API that demonstrates how to use my system.",
      "version": "1.0.0"
    },
...
    "servers": [
        {
            "url": "/"
        }
    ],
    "components": {
        "schemas": {
            "Products":{
                "type": "object",
                "properties": {
                    "product_id": {
                        "type": "string"
                    },
                    "product_name": {
                        "type": "string"
                    },
                    "product_price": {
                        "type": "integer"
                    },
                    "product_description": {
                        "type": "string"
                    }
                }
            },
            "Id": {
                "type": "integer",
                "properties": {
                    "name" : {
                        "type": "integer"
                    }
                }
            },
            "Runs": {
                "type": "object",
                "properties": {
                    "product_id": {
                        "type": "integer"
                    },
                    "flour_kneaded": {
                        "type": "integer"
                    },
                    "oil_used": {
                        "type": "integer"
                    },
                    "packets_produced": {
                        "type": "integer"
                    }
                }
            },
            "Ingredients": {
                "type": "object",
                "properties": {
                    "ingredient_id": {
                        "type": "string"
                    },
                    "ingredient_name": {
                        "type": "string"
                    },
                    "ingredient_quantity": {
                        "type": "integer"
                    },
                    "ingredient_measurement": {
                        "type": "string"
                    },
                    "ingredient_cost": {
                        "type": "integer"
                    }
                }
            },
            "Recipes": {
                "type": "object",
                "properties": {
                    "product_id": {
                        "type": "integer"
                    },
                    "description": {
                        "type": "string"
                    },
                    "yield_amount": {
                        "type": "integer"
                    }
                }
            },
            "RecipeDetails": {
                "type": "object",
                "properties": {
                    "recipe_id": {
                        "type": "integer"
                    },
                    "ingredient_id": {
                        "type": "integer"
                    },
                    "quantity": {
                        "type": "integer"
                    },
                    "measurement": {
                        "type": "string"
                    },
                    "unit_cost": {
                        "type": "integer"
                    }
                }
            },
...
        }
    },
    "paths": {
        "/production/products": {
          "get": {
            "tags": [
                "Get all products from API"
            ],
            "summary": "Get all products",
            "responses": {
              "200": {
                "description": "Successful operation",
                "schema": {
                    "$ref": "#/components/schemas/Products"
                }
              },
              "404": {
                "description": "NOT FOUND",
                "schema": {
                    "$ref": "#/components/schemas/Products"
                    }
                }
          }
        },
          "post": {
            "tags": [
                "Add a product to API"
            ],
            "summary": "Add product",
            "requestBody": {
                "description": "Post new product to database",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Products"
                        }
                    }
                }
            },
            "responses": {
                "201": {
                    "description": "OK",
                    "schema": {
                        "$ref": "#/components/schemas/Products"
                    }
                },
                "404": {
                    "description": "NOT FOUND",
                    "schema": {
                        "$ref": "#/components/schemas/Products"
                    }
                }
            }
          }
        },
        "/production/products/{id}": {
            "get": {
                "tags": [
                    "Get product from API"
                ],
                "summary": "Get Product",
                "parameters": [{
                    "name": "id",
                    "in": "path",
                    "description": "product id to update",
                    "required": true,
                    "type": "integer",
                    "schema": {
                        "$ref": "#/components/schemas/Id"
                    }
                }],

                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/components/schemas/Products"
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "schema": {
                            "$ref": "#/components/schemas/Products"
                        }
                    }
                }
            },

            "put": {
                "tags": [
                    "Update Product on API"
                ],
                "parameters": [{
                    "name": "id",
                    "in": "path",
                    "description": "product id to update",
                    "required": true,
                    "type": "integer",
                    "schema": {
                        "$ref": "#/components/schemas/Id"
                    }
                }],

                "summary": "Update Product",
                "requestBody": {
                    "description": "Update Existing product on Database",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Products"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/components/schemas/Products"
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "schema": {
                            "$ref": "#/components/schemas/Products"
                        }
                    }
                } 
            },
            "delete": {
                "tags": [
                    "Delete product on API"
                ],
                "parameters": [{
                    "name": "id",
                    "in": "path",
                    "description": "product id to delete",
                    "required": true,
                    "type": "integer",
                    "schema": {
                        "$ref": "#/components/schemas/Id"
                    }
                }],
                "summary": "Delete Product",

                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/components/schemas/Products"
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "schema": {
                            "$ref": "#/components/schemas/Products"
                        }
                    }
                }
            }
        },
    "/production/productionruns": {
        "get": {
            "tags": [
                "Get all production runs from API"
            ],
            "summary": "Get all production runs",
            "responses": {
            "200": {
                "description": "Successful operation",
                "schema": {
                    "$ref": "#/components/schemas/Runs"
                }
            },
            "404": {
                "description": "NOT FOUND",
                "schema": {
                    "$ref": "#/components/schemas/Runs"
                    }
                }
        }
        },
        "post": {
            "tags": [
                "Add a production run to API"
            ],
            "summary": "Add production run",
            "requestBody": {
                "description": "Post new production run to database",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Runs"
                        }
                    }
                }
            },
            "responses": {
                "201": {
                    "description": "OK",
                    "schema": {
                        "$ref": "#/components/schemas/Runs"
                    }
                },
                "404": {
                    "description": "NOT FOUND",
                    "schema": {
                        "$ref": "#/components/schemas/Runs"
                    }
                }
            }
        }
        },
        "/production/productionruns/{id}": {
            "get": {
                "tags": [
                    "Get production run from API"
                ],
                "summary": "Get Production Run",
                "parameters": [{
                    "name": "id",
                    "in": "path",
                    "description": "production run id to update",
                    "required": true,
                    "type": "integer",
                    "schema": {
                        "$ref": "#/components/schemas/Id"
                    }
                }],

                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/components/schemas/Runs"
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "schema": {
                            "$ref": "#/components/schemas/Runs"
                        }
                    }
                }
            },

            "put": {
                "tags": [
                    "Update Production Run on API"
                ],
                "parameters": [{
                    "name": "id",
                    "in": "path",
                    "description": "production run id to update",
                    "required": true,
                    "type": "integer",
                    "schema": {
                        "$ref": "#/components/schemas/Id"
                    }
                }],

                "summary": "Update Production Run",
                "requestBody": {
                    "description": "Update Existing production run on Database",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Runs"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/components/schemas/Runs"
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "schema": {
                            "$ref": "#/components/schemas/Runs"
                        }
                    }
                } 
            },
            "delete": {
                "tags": [
                    "Delete production run on API"
                ],
                "parameters": [{
                    "name": "id",
                    "in": "path",
                    "description": "production run id to delete",
                    "required": true,
                    "type": "integer",
                    "schema": {
                        "$ref": "#/components/schemas/Id"
                    }
                }],
                "summary": "Delete Production Run",

                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/components/schemas/Runs"
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "schema": {
                            "$ref": "#/components/schemas/Runs"
                        }
                    }
                }
            }
        },
        "/production/ingredients": {
            "get": {
                "tags": [
                    "Get all ingredients from API"
                ],
                "summary": "Get all ingredients",
                "responses": {
                "200": {
                    "description": "Successful operation",
                    "schema": {
                        "$ref": "#/components/schemas/Ingredients"
                    }
                },
                "404": {
                    "description": "NOT FOUND",
                    "schema": {
                        "$ref": "#/components/schemas/Ingredients"
                        }
                    }
            }
            },
            "post": {
                "tags": [
                    "Add an ingredient to API"
                ],
                "summary": "Add ingredient",
                "requestBody": {
                    "description": "Post new ingredient to database",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Ingredients"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/components/schemas/Ingredients"
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "schema": {
                            "$ref": "#/components/schemas/Ingredients"
                        }
                    }
                }
            }
            },
        "/production/ingredients/{id}": {
            "get": {
                "tags": [
                    "Get ingredient from API"
                ],
                "summary": "Get Ingredient",
                "parameters": [{
                    "name": "id",
                    "in": "path",
                    "description": "ingredient id to update",
                    "required": true,
                    "type": "integer",
                    "schema": {
                        "$ref": "#/components/schemas/Id"
                    }
                }],

                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/components/schemas/Ingredients"
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "schema": {
                            "$ref": "#/components/schemas/Ingredients"
                        }
                    }
                }
            },

            "put": {
                "tags": [
                    "Update Ingredient on API"
                ],
                "parameters": [{
                    "name": "id",
                    "in": "path",
                    "description": "ingredient id to update",
                    "required": true,
                    "type": "integer",
                    "schema": {
                        "$ref": "#/components/schemas/Id"
                    }
                }],

                "summary": "Update Ingredient",
                "requestBody": {
                    "description": "Update Existing ingredient on Database",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Ingredients"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/components/schemas/Ingredients"
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "schema": {
                            "$ref": "#/components/schemas/Ingredients"
                        }
                    }
                } 
            },
            "delete": {
                "tags": [
                    "Delete ingredient on API"
                ],
                "parameters": [{
                    "name": "id",
                    "in": "path",
                    "description": "ingredient id to delete",
                    "required": true,
                    "type": "integer",
                    "schema": {
                        "$ref": "#/components/schemas/Id"
                    }
                }],
                "summary": "Delete Product",

                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/components/schemas/Ingredients"
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "schema": {
                            "$ref": "#/components/schemas/Ingredients"
                        }
                    }
                }
            }
        },
        "/production/recipes": {
            "get": {
                "tags": [
                    "Get all recipes from API"
                ],
                "summary": "Get all recipes runs",
                "responses": {
                "200": {
                    "description": "Successful operation",
                    "schema": {
                        "$ref": "#/components/schemas/Recipes"
                    }
                },
                "404": {
                    "description": "NOT FOUND",
                    "schema": {
                        "$ref": "#/components/schemas/Recipes"
                        }
                    }
            }
            },
            "post": {
                "tags": [
                    "Add a recipe to API"
                ],
                "summary": "Add recipe",
                "requestBody": {
                    "description": "Post new recipe to database",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Recipes"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/components/schemas/Recipes"
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "schema": {
                            "$ref": "#/components/schemas/Recipes"
                        }
                    }
                }
            }
            },
...
        "/production/recipedetails": {
            "get": {
                "tags": [
                    "Get all recipe details from API"
                ],
                "summary": "Get all recipe details",
                "responses": {
                "200": {
                    "description": "Successful operation",
                    "schema": {
                        "$ref": "#/components/schemas/RecipeDetails"
                    }
                },
                "404": {
                    "description": "NOT FOUND",
                    "schema": {
                        "$ref": "#/components/schemas/RecipeDetails"
                        }
                    }
            }
            },
            "post": {
                "tags": [
                    "Add a recipe detail to API"
                ],
                "summary": "Add recipe detail",
                "requestBody": {
                    "description": "Post new recipe detail to database",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RecipeDetails"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/components/schemas/RecipeDetails"
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "schema": {
                            "$ref": "#/components/schemas/RecipeDetails"
                        }
                    }
                }
            }
            },
...
    }
}